refactor(rust): extract provekit-linker library crate (LSP+linker step 1)
Pure refactor. Creates implementations/rust/provekit-linker as a lib-only
workspace crate exposing:
- LinkerInputs / LinkerOutput / KitContract / KitCallEdge / LinkerError
  (all Serialize+Deserialize so the daemon can reconstruct from parseFile streams)
- link(LinkerInputs) -> LinkerOutput — pure, no global state, no I/O

LinkerError gains a `file: Option<String>` field populated from the
call-edge's callSiteLocus.file for per-file LSP diagnostic mapping
(daemon step 2 requires this per spec #126 §3).

provekit-cli is updated:
- cmd_link.rs is now a thin I/O wrapper: gathers lifted contracts and
  call-edges, calls provekit_linker::link(), writes bundle to disk.
- lib.rs (provekit_cli_test_support shim with a third copy of the algo)
  is deleted.
- tests/polyglot_smoke.rs imports from provekit_linker directly.

Byte-identity gate: both fixture CIDs are pinned in a new unit test in
provekit-linker and verified to match the #124 baseline exactly.

Closes LSP+linker step 1. Daemon (step 2) can now depend on this crate.
